+++ app/src/main/java/com/rave/simplemath/viewmodel/SumViewModel.kt
@@ -30,7 +30,7 @@ class SumViewModel : ViewModel() {
         _sum.update { it.copy(isLoading = true) }
         val expression = "$firstArg%2B$secondArg"
         val sumResponse = MathRepo.evaluateExpression(expression, Dispatchers.IO)
-        _sum.update { it.copy(isLoading = false, sum = sumResponse.toInt()) }
+        _sum.update { it.copy(isLoading = false, sum = sumResponse) }
     }
 }
